Refactor: 버텀 탭 컴포넌트 분리
